pyproject.toml
[project]
name = "compare-rags"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.3.2",
    "llama-index>=0.12.23",
    "llama-index-embeddings-huggingface>=0.5.2",
    "llama-index-llms-mistralai>=0.4.0",
    "neo4j>=5.28.1",
    "neo4j-graphrag[mistralai]>=1.4.0",
    "python-dotenv>=1.0.1",
    "transformers>=4.49.0",
]
